// client/types/Employee.ts
export default interface Employee {
  _id: string;
  firstName: string;
  lastName: string;
  employeeID: string;
  dateBirth: Date;
  role: string;
  department: string;
  dateHired: Date;
  salary: string;
  status: string;
  telephone: string;
  address: string;
  emergencyContact: string;
  relationship: string;
  contactPhone: string;
  remainingLeave: number;
}